Convert App.test to TS and update snapshot
